requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
requests==2.32.3
PyJWT[crypto]==2.10.1
python-dotenv==1.0.1
pydantic-settings==2.7.0
